Fastapi/requirements1.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==0.7.16
albucore==0.0.23
albumentations==2.0.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
atomicwrites==1.4.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.5
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
curio==1.6
cvzone==1.6.1
cycler==0.12.1
Cython==3.0.12
datasets==3.2.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docrepr==0.2.0
docutils==0.21.2
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
image-classifiers==1.0.0
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipyparallel==9.0.0
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
langchain-core==0.3.59
langchain-google-genai==2.0.10
langsmith==0.3.42
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
mask-rcnn==2.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mediapipe==0.10.21
MedPy==0.5.2
mistune==3.1.1
ml-collections==0.1.1
ml_dtypes==0.5.1
monai==1.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mrcnn==0.2
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
narwhals==1.38.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-contrib-python-headless==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patchify==0.2.3
pathspec==0.12.1
peft==0.15.2
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==3.1.1
pretrainedmodels==0.7.4
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==6.1.0
pure_eval==0.2.3
py==1.11.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==6.2.5
pytest-asyncio==0.20.3
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytweening==1.2.0
pytz==2024.2
pyvista==0.44.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
scooby==0.10.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
segmentation-models==1.0.1
segmentation_models_pytorch==0.4.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shapely==2.0.7
SimpleITK==2.4.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
Sphinx==7.4.7
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
streamlit==1.45.0
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.22.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.28.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
vtk==9.3.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
